LayoutTests/crypto/webkitSubtle/wrapKey-check-usage-expected.txt

CONSOLE MESSAGE: line 28: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
CONSOLE MESSAGE: line 1: Key usages do not include 'wrapKey'
Verify that usage is respected in wrapKey operation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing an RSA key...
Importing an AES key...
Wrapping it...
PASS crypto.webkitSubtle.wrapKey("raw", aesKey, publicKey, wrapAlgorithm) threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS successfullyParsed is true

TEST COMPLETE
